Try to circumvent JET issue
Previously, we had the issue that `AB_product_storage` could either be a
reference to a matrix, or a new matrix. This made it challenging for the
compiler to reason about the type of the variable.
We try to solve this by introducing a new "function barrier", i.e.,
putting the performance critical code inside a new function so that
objects' type can be properly infered.