Add some @zero_derivatives for logging (#815)
* Mark Base.fixup_stdlib_path as having @zero_derivative

This comes up because in Julia 1.12 logmsg_code gets the folder of
Core.Compiler if needed to resolve paths, which results in a ccall
with no known rule. See
https://github.com/JuliaLang/julia/pull/57274.

* Make Base.CoreLogging.handle_message_nothrow as having @zero_derivative

This comes up in logmsg_code, because in Julia v1.12
handle_message_nothrow got a try-catch structure, which creates an
UpsilonNode that Mooncake can't handle.